+++ eslint.config.js
@@ -1,0 +1,6 @@
+"use strict";
+
+module.exports = [
+  { ignores: [ "test/data/**/*" ] },
+  ...require("./index.js"),
+];