+++ test/feature/lint-typescript-feature.js
@@ -7,8 +7,7 @@ Feature("linting ts files", () => {
     let eslint;
     Given("we have an eslint instance with the base config", () => {
       eslint = new ESLint({
-        overrideConfigFile: "typescript.js",
-        useEslintrc: false,
+        overrideConfigFile: "ts.js",
         ignore: false,
       });
     });
@@ -31,8 +30,7 @@ Feature("linting ts files", () => {
     let eslint;
     Given("we have an eslint instance with the base config", () => {
       eslint = new ESLint({
-        overrideConfigFile: "typescript.js",
-        useEslintrc: false,
+        overrideConfigFile: "ts.js",
         ignore: false,
       });
     });
@@ -47,7 +45,7 @@ Feature("linting ts files", () => {
     });
 
     And("one file should have errors", () => {
-      expect(results[0].messages.length).to.be.gt(1);
+      expect(results[0].messages.length).to.eql(1);
     });
 
     And("one error should be because it has enum", () => {
@@ -62,8 +60,7 @@ Feature("linting ts files", () => {
     let eslint;
     Given("we have an eslint instance with the base config", () => {
       eslint = new ESLint({
-        overrideConfigFile: "typescript.js",
-        useEslintrc: false,
+        overrideConfigFile: "ts.js",
         ignore: false,
       });
     });
@@ -93,8 +90,7 @@ Feature("linting ts files", () => {
     let eslint;
     Given("we have an eslint instance with the base config", () => {
       eslint = new ESLint({
-        overrideConfigFile: "typescript.js",
-        useEslintrc: false,
+        overrideConfigFile: "ts.js",
         ignore: false,
       });
     });
@@ -124,8 +120,7 @@ Feature("linting ts files", () => {
     let eslint;
     Given("we have an eslint instance with the base config", () => {
       eslint = new ESLint({
-        overrideConfigFile: "typescript.js",
-        useEslintrc: false,
+        overrideConfigFile: "ts.js",
         ignore: false,
       });
     });
@@ -155,8 +150,7 @@ Feature("linting ts files", () => {
     let eslint;
     Given("we have an eslint instance with the base config", () => {
       eslint = new ESLint({
-        overrideConfigFile: "typescript.js",
-        useEslintrc: false,
+        overrideConfigFile: "ts.js",
         ignore: false,
       });
     });